requirements.txt
streamlit
Pillow
google-auth
google-auth-httplib2
google-api-python-client